requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==3.1
importlib-metadata==4.3.1
ipfshttpclient==0.6.1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.17.1
pycryptodome==3.10.1
pypiwin32==223
pyrsistent==0.17.3
python-dotenv==0.17.1
pytz==2021.1
pywin32==300
requests==2.25.1
rlp==2.0.1
six==1.16.0
toolz==0.11.1
transliterate==1.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
varint==1.0.2
web3==5.12.2
websockets==8.1
yarl==1.6.3
zipp==3.4.1
